requirements.txt
absl-py==1.4.0
add-parent-path==0.1.11
altair==4.2.0
annotated-types==0.5.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
async-generator==1.10
atoma==0.0.17
attrs==22.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.5
boto3==1.28.22
botocore==1.31.22
branca==0.6.0
Brotli==1.0.9
cachetools==5.2.1
CatastroInspireDownloader==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1670575068857/work
contourpy==1.0.7
cryptography==39.0.1
cssselect2==0.7.0
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.2.4
dm-tree==0.1.8
docopt==0.6.2
duckdb==0.8.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et-xmlfile==1.1.0
etils==1.0.0
exceptiongroup==1.1.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastjsonschema==2.16.3
filelock==3.12.0
Fiona==1.8.22
flatbuffers==23.1.21
folium==0.14.0
fonttools==4.38.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.3.5
fuzzywuzzy==0.18.0
gast==0.4.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.30
gobject==0.1.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
gspread==5.7.2
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httplib2==0.21.0
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel @ file:///D:/bld/ipykernel_1673894762113/work
ipytablewidgets==0.3.1
ipython @ file:///D:/bld/ipython_1672758759297/work
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///D:/bld/jupyter_core_1674530221624/work
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kanilog==0.3.3
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
logzero==1.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mistune==2.0.5
multitasking==0.0.11
munch==2.5.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy==0.5.3
pdfkit==1.0.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.4.0
plotly==5.14.0
plotly-express==0.4.1
pluggy==1.0.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==4.21.12
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
psycopg2==2.9.7
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
pydeck==0.8.0
PyDrive==1.3.1
pydyf==0.7.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pygwalker==0.3.3
pykml==0.2.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyphen==0.14.0
pyproj==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
reportlab==4.0.4
requests==2.22.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.2.0
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.0
semver==2.13.0
Send2Trash==1.8.2
shapely==2.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlglot==17.14.2
sqlparse==0.4.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.13.5
stdlogging==0.16
streamlit==1.17.0
streamlit-folium==0.11.1
streamlit-metrics==0.1.0
streamlit-option-menu==0.3.6
streamlit-state==0.1.0
TA-Lib @ file:///C:/Users/elope/OneDrive/Documentos/trading/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1673359992537/work
traittypes==0.2.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
unicode==2.9
Unidecode==1.3.6
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.25.11
validators==0.20.0
vega==4.0.0
virtualenv==20.22.0
watchdog==2.2.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
weasyprint==59.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
yfinance==0.2.12
zipp==3.11.0
zopfli==0.2.2
